Added hierarchy of errors in separate module
