+++ src/ultimate_notion/__init__.py
@@ -49,15 +49,15 @@ from ultimate_notion.blocks import (
     Video,
 )
 from ultimate_notion.core import get_active_session
-from ultimate_notion.database import Database, ReadOnlyPropertyError
+from ultimate_notion.database import Database
 from ultimate_notion.file import Emoji, FileInfo
 from ultimate_notion.obj_api.enums import AggFunc, BGColor, CodeLang, Color, NumberFormat, VState
 from ultimate_notion.option import Option, OptionNS
 from ultimate_notion.page import Page
 from ultimate_notion.query import Condition, prop
 from ultimate_notion.rich_text import join, math, mention, text
-from ultimate_notion.schema import Property, PropType, Schema, SchemaError, SelfRef
-from ultimate_notion.session import Session, SessionError, UnknownPageError, UnknownUserError
+from ultimate_notion.schema import Property, PropType, Schema, SelfRef
+from ultimate_notion.session import Session
 from ultimate_notion.user import User
 from ultimate_notion.utils import SList
 
@@ -98,21 +98,16 @@ __all__ = [
     'PropType',
     'Property',
     'Quote',
-    'ReadOnlyPropertyError',
     'SList',
     'Schema',
-    'SchemaError',
     'SelfRef',
     'Session',
-    'SessionError',
     'SyncedBlock',
     'Table',
     'TableOfContents',
     'TableRow',
     'ToDoItem',
     'ToggleItem',
-    'UnknownPageError',
-    'UnknownUserError',
     'User',
     'VState',
     'Video',
