+++ src/ultimate_notion/utils.py
@@ -19,16 +19,9 @@ from packaging.version import Version
 from pydantic import BaseModel
 
 from ultimate_notion import __version__
+from ultimate_notion.errors import EmptyListError, MultipleItemsError
 
 T = TypeVar('T')  # ToDo: Use new syntax when requires-python >= 3.12
-
-
-class EmptyListError(Exception):
-    """Custom exception for an empty list in SList."""
-
-
-class MultipleItemsError(Exception):
-    """Custom exception for a list with multiple items in SList."""
 
 
 class SList(list[T]):
