+++ tests/test_session.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 import pytest
 
 import ultimate_notion as uno
+from ultimate_notion.errors import UnknownPageError, UnknownUserError
 
 from .conftest import CONTACTS_DB
 
@@ -31,7 +32,7 @@ def test_whoami_get_user(notion: uno.Session):
     assert user.id == me.id
     user = notion.get_user(me.id, use_cache=False, raise_on_unknown=False)
     assert user.name == 'Github Unittests'
-    with pytest.raises(uno.UnknownUserError):
+    with pytest.raises(UnknownUserError):
         unknown_id = 'f3f2e850-b5d4-11ef-ac7e-96584d5248b2'
         notion.get_user(unknown_id, use_cache=False)
 
@@ -41,7 +42,7 @@ def test_get_page_by_id(notion: uno.Session, intro_page: uno.Page):
     page_by_id = notion.get_page(intro_page.id)
     assert page_by_id.title == 'Getting Started'
     assert page_by_id == intro_page
-    with pytest.raises(uno.UnknownPageError):
+    with pytest.raises(UnknownPageError):
         unknown_id = 'f3f2e850-b5d4-11ef-ac7e-96584d5248b2'
         notion.get_page(unknown_id)
